few more funcitons

+++ s1_cslc_parallel.py
@@ -13,7 +13,7 @@ import yaml
 import argparse
 import time
 import pandas as pd
-import shutil
+import datetime
 
 from compass.utils import iono
 dict_pol_to_load = {
@@ -27,6 +27,20 @@ dict_pol_to_load = {
     '1SHV':'HV',
 }
 
+
+def get_burst_id_list(burst_csv_path):
+    '''
+    Extract burst lists to process from 
+
+    
+    '''
+    df_csv_burst_id = pd.read_csv(burst_csv_path)
+    burst_id_list = []
+
+    for row in df_csv_burst_id.iterrows():
+        burst_id_list.append(row[1]['burst_id_jpl'])
+
+    return burst_id_list
 
 
 def prepare_batch_process(src_csv_path, project_dir,
@@ -177,23 +191,47 @@ def get_all_burst_id(path_safe):
 
 def find_common_bursts(list_safe_zip):
     '''
-    Docstring here
+    Find the common burst IDs in the input list of SAFE file
+  
+    Parameter
+    ---------
+    list_safe_zip: list
+        List of SAFE files.
+
+    Returns
+    -------
+    set_common_burst:
+        Common bursts in the SAFE file
     '''
+
     num_safe_zip = len(list_safe_zip)
 
-    # initial set
-    set_common_burst = get_all_burst_id(list_safe_zip[0])
+    # Extract the burst IDs in each acquisition dates
+    burst_id_by_daq_dict = {}
+    for i_safe, safe_name in enumerate(list_safe_zip):
+        print(f'Analyzing bursts in a SAFE file: {i_safe} / {num_safe_zip}')
+        safe_id = os.path.basename(safe_name).rstrip('.zip')
+        sensing_start_date = safe_id.split('_')[5]
 
-    for i_safe in range(1, num_safe_zip):
-        print(f'Taking a look {i_safe} / {num_safe_zip}')
-        set_burst_id_safe = get_all_burst_id(list_safe_zip[i_safe])
-        set_common_burst = set_common_burst & set_burst_id_safe
+        if not sensing_start_date in burst_id_by_daq_dict.keys():
+            burst_id_by_daq_dict[sensing_start_date] = set([])
 
-    return list(set_common_burst)
+        set_common_burst = get_all_burst_id(list_safe_zip[0])
+        burst_id_by_daq_dict[sensing_start_date] =\
+            burst_id_by_daq_dict[sensing_start_date].union(set_common_burst)
+
+    # Set up the initial set for common burst
+    common_burst_set = burst_id_by_daq_dict[sensing_start_date]
+
+    for _, set_burst_id in burst_id_by_daq_dict.items():
+        common_burst_set = common_burst_set & set_burst_id
+
+    return list(common_burst_set)
 
 
-def spawn_runconfig(ref_runconfig_path, safe_dir, orbit_dir):
+def spawn_runconfig_old(ref_runconfig_path, safe_dir, orbit_dir):
     '''
+    OLD version of `spawn_runconfig`
     Split the input runconfig into single burst runconfigs.
     Writes out the runconfigs.
     Return the list of the burst runconfigs.
@@ -253,6 +291,88 @@ def spawn_runconfig(ref_runconfig_path, safe_dir, orbit_dir):
     return runconfig_burst_list
 
 
+def spawn_runconfig(ref_runconfig_path, df_csv, project_dir, burst_id_csv_path=None):
+    '''
+    Split the input runconfig into single burst runconfigs.
+    Writes out the runconfigs.
+    Return the list of the burst runconfigs.
+
+    Parameters:
+
+    Returns:
+    list_runconfig_burst: list(str)
+        List of the burst runconfigs
+    list_logfile_burst: list(str)
+        List of the burst logfiles,
+        which corresponds to `list_runconfig_burst`
+    '''
+
+    if burst_id_csv_path:
+        candidate_burst_ids = set(get_burst_id_list(burst_id_csv_path))
+    else:
+        candidate_burst_ids = None
+
+    with open(ref_runconfig_path, 'r+', encoding='utf8') as fin:
+        runconfig_dict_ref = yaml.safe_load(fin.read())
+
+    scratch_dir_base = runconfig_dict_ref['runconfig']['groups']['product_path_group']['scratch_path']
+    burst_id_list_ref = runconfig_dict_ref['runconfig']['groups']['input_file_group']['burst_id']
+
+    os.makedirs(scratch_dir_base, exist_ok=True)
+
+
+    flag_tec_file_availale = 'TEC file' in df_csv.columns
+
+    runconfig_burst_list = []
+
+    for row in df_csv.iterrows():
+        safe_path = row[1]['DOWNLOADED SAFE']
+        #if safe_path.startswith(project_dir):
+        #    safe_path = safe_path.lstrip(project_dir)
+        orbit_path = row[1]['Orbit path']
+        #if orbit_path.startswith(project_dir):
+        #    safe_path = orbit_path.lstrip(project_dir)
+        if flag_tec_file_availale:
+            tec_path = row[1]['TEC file']
+        #    if tec_path.startswith(project_dir):
+        #        tec_path = tec_path.lstrip(project_dir)
+        else:
+            tec_path = ''
+
+        start_time = datetime.datetime.fromisoformat(row[1]['Start Time'])
+        str_start_time = start_time.strftime('%Y%m%d_%H%M%S')
+        
+        # Determine the burst ID to process
+        if burst_id_list_ref:
+            burst_id_list = burst_id_list_ref
+        else:
+            all_burst_in_zip = get_all_burst_id(safe_path)
+            if candidate_burst_ids:
+                burst_id_list = list(all_burst_in_zip & candidate_burst_ids)
+                if len(burst_id_list) == 0:
+                    continue
+        
+        for burst_id in burst_id_list:
+            runconfig_path = os.path.join(scratch_dir_base, f'runconfig_{str_start_time}_{burst_id}.yaml')
+            scratch_dir = os.path.join(scratch_dir_base, f'temp_{str_start_time}_{burst_id}')
+            runconfig_dict_out = runconfig_dict_ref.copy()
+
+            runconfig_dict_out['runconfig']['groups']['input_file_group']['safe_file_path'] = [safe_path]
+            runconfig_dict_out['runconfig']['groups']['input_file_group']['orbit_file_path'] = [orbit_path]
+            runconfig_dict_out['runconfig']['groups']['product_path_group']['scratch_path'] = scratch_dir
+            runconfig_dict_out['runconfig']['groups']['input_file_group']['burst_id'] = [burst_id]
+            runconfig_dict_out['runconfig']['groups']['dynamic_ancillary_file_group']['tec_file'] = tec_path
+
+            runconfig_burst_list.append(runconfig_path)
+
+            with open(runconfig_path, 'w+', encoding='utf8') as fout:
+                yaml.dump(runconfig_dict_out, fout)
+    
+    return runconfig_burst_list
+
+
+
+
 def get_parser():
     '''Initialize YamlArgparse class and parse CLI arguments for OPERA RTC.
     Modified after copied from `rtc_s1.py`
@@ -262,18 +382,29 @@ def get_parser():
     parser.add_argument('run_config_path',
                         type=str,
                         default=None,
-                        help='Path to run config file')
+                        help='Path to reference run config file')
 
-    parser.add_argument('dir_safe',
+    parser.add_argument('csv_path',
                         type=str,
                         default=None,
-                        help='Directory for the safe file')
-
-    parser.add_argument('-o',
-                        dest='dir_orbit',
+                        help='ASF VERTEX search result as .csv')
+    
+    parser.add_argument('csv_out',
                         type=str,
-                        default='orbits',
-                        help='Directory for orbit')
+                        default=None,
+                        help='Stack processing preparation result')
+
+    parser.add_argument('-p',
+                        dest='project_dir',
+                        type=str,
+                        default=os.getcwd(),
+                        help='Project directory')
+
+    parser.add_argument('-b',
+                        dest='candidate_burst_path',
+                        type=str,
+                        default=os.getcwd(),
+                        help='candidate burst IDs in the column "burst_id_jpl" as .csv file')
 
     # Determine the default # of concurrent workers
     ncpu_default = os.cpu_count()
@@ -286,18 +417,6 @@ def get_parser():
                         type=int,
                         default=ncpu_default,
                         help='Number of concurrent workers.')
-
-    parser.add_argument('--log',
-                        '--log-file',
-                        dest='log_file',
-                        type=str,
-                        help='Log file')
-
-    parser.add_argument('--full-log-format',
-                        dest='full_log_formatting',
-                        action='store_true',
-                        default=False,
-                        help='Enable full formatting of log messages')
 
     return parser
 
@@ -322,11 +441,11 @@ def process_runconfig(path_runconfig_burst):
     rtnval = subprocess.run(list_arg_subprocess)
 
     # TODO Add some routine to take a look into `rtnval` to see if everything is okay.
+    if rtnval.returncode == 0:
+        os.remove(path_runconfig_burst)
 
-    os.remove(path_runconfig_burst)
 
-
-def process_frame_parallel(arg_in):
+def run(arg_in):
     '''
     Take in the parsed arguments from CLI,
     split the original runconfign into bursts,
@@ -337,8 +456,13 @@ def process_frame_parallel(arg_in):
         Parsed argument. See `get_rtc_s1_parser()`
     '''
 
+    prepare_batch_process(arg_in.csv_path, arg_in.project_dir, arg_in.csv_out)
+    
+    df_csv = pd.read_csv(arg_in.csv_out)
+    
     t0 = time.time()
-    list_burst_runconfig = spawn_runconfig(arg_in)
+    #list_burst_runconfig = spawn_runconfig(arg_in)
+    list_burst_runconfig = spawn_runconfig(arg_in.run_config_path, df_csv, arg_in.project_dir, arg_in.candidate_burst_path)
 
     with multiprocessing.Pool(arg_in.num_workers) as p:
         p.map(process_runconfig, list_burst_runconfig)
@@ -352,6 +476,6 @@ if __name__=='__main__':
     parser = get_parser()
     args = parser.parse_args()
 
-    process_frame_parallel(args)
+    run(args)
     #spawn_runconfig(args)
